requirements.txt
async-generator
attrs
certifi
chardet
defusedxml
dj-database-url
Django
django-allauth
django-cors-headers
dj_rest_auth
djangorestframework
gunicorn
h11
idna
oauthlib
outcome
psycopg2
PySocks
python-dotenv
python3-openid
pytz
requests-oauthlib
selenium
six
sniffio
sortedcontainers
trio
trio-websocket
whitenoise
wsproto
